+++ ScatterPointBrush.h
@@ -17,7 +17,7 @@ public:
 	void BrushBegin(const Point source, const Point target) override;
 	void BrushMove(const Point source, const Point target) override;
 
-	int randomCountFactor = 5;
+	float randomCountFactor = 5;
 };
 
 #endif
